backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
groq>=0.11
pydantic>=2.5
pypdf>=4.0
python-multipart>=0.0.9
